requirements.txt
# Core dependencies
pdfplumber==0.11.0
requests>=2.32.5
pydantic>=2.11.0
pandas>=2.2.0
google-generativeai>=0.8.0
openai>=1.0.0
dspy-ai>=2.4.0

# OCR dependencies
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# Beautiful logging and CLI
rich>=13.7.0

# Jupyter
jupyter==1.0.0
notebook>=7.0.6
ipykernel>=6.27.1

# Optional but recommended
python-dotenv==1.0.0
